test: moves unit tests for ObjectBuilder and DTOFactory functionality to according directories
